Remove HashSetWildcardPool and its usage in CmdletContext
The `CmdletContext.cs` file had a line removed that added a pool for `HashSet<Wildcard>` using `HashSetWildcardPool`. This line was:
    AddPool<HashSet<Wildcard>, HashSetWildcardPool>(services);

The `HashSetWildcardPool.cs` file was entirely removed. This file contained the implementation of the `HashSetWildcardPool` class, which was a custom object pool for `HashSet<Wildcard>`. The class included methods for constructing, resetting, and managing the capacity of `HashSet<Wildcard>` objects.

These changes suggest that the pooling mechanism for `HashSet<Wildcard>` objects is no longer needed or has been replaced by a different implementation.