Add dynamic variable resolution and config improvements
Introduced a `resolveVariables` function in `Debug.ps1` to
dynamically resolve placeholders in strings using PowerShell
variables. Updated `Connect-SonarrInstance` to use this
function for resolving `Url` and `ApiKey` values, enabling
placeholders in `debugging.json`.

Replaced the `ApiKey` in `debugging.json` with a placeholder
`{skey}` for dynamic resolution. Ensured `$BuildDependenciesJson`
is converted to a fully qualified path if necessary. Improved
file handling in the `Get-ChildItem` pipeline for the `PSTypes`
directory.